iterator in a stu
